move duplicate to service
